Better detection of mounted sshfs filesystems.
We used to look for /usr/bin/sshfs process instances
to see if something is mounted, but that's not very realiable,
because it may tell you that something is mounted
even though it hasn't fully completed the mounting process.

This new method uses mount -l, which is more reliable, faster and simpler.