Fix mounted systems detection for ids containing dots or dashes
According to the current validation rules, system ids be
alphanumeric and additionally contain dots and dashes.

Detection of mounted systems, however, only worke with alphanumeric
system ids and failed to detect other systems as mounted.

This caused mounting to work for them, but a successful mount
was never reported as it couldn't be detected, giving out
"Mounting failed for a reason unknown to sftpman." errors.